@@ -12,7 +12,7 @@ import {
     untracked,
     ViewChild
 } from '@angular/core';
-import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+import { takeUntilDestroyed, toSignal } from '@angular/core/rxjs-interop';
 import { FormControl, ReactiveFormsModule } from '@angular/forms';
 
 import { MenuItem, MessageService } from 'primeng/api';
@@ -26,7 +26,7 @@ import { OverlayPanelModule } from 'primeng/overlaypanel';
 import { PaginatorModule, PaginatorState } from 'primeng/paginator';
 import { SkeletonModule } from 'primeng/skeleton';
 
-import { debounceTime, distinctUntilChanged } from 'rxjs/operators';
+import { debounceTime, distinctUntilChanged, filter, map } from 'rxjs/operators';
 
 import {
     DotESContentService,
@@ -39,15 +39,17 @@ import { DotMessagePipe } from '@dotcms/ui';
 import { DotPaletteListStore } from './store/store';
 
 import {
+    DotPaletteSearchParams,
     DotPaletteSortOption,
     DotPaletteViewMode,
     DotUVEPaletteListTypes,
     DotUVEPaletteListView
 } from '../../models';
 import {
+    buildPaletteMenuItems,
+    EMPTY_MESSAGE_CONTENTLETS,
     EMPTY_MESSAGE_SEARCH,
     EMPTY_MESSAGES,
-    getSortActiveClass,
     LOADING_ROWS_MOCK
 } from '../../utils';
 import { DotFavoriteSelectorComponent } from '../dot-favorite-selector/dot-favorite-selector.component';
@@ -63,8 +65,8 @@ import { DotUVEPaletteContenttypeComponent } from '../dot-uve-palette-contenttyp
  * <dot-uve-palette-list
  *   [type]="'content'"
  *   [languageId]="1"
- *   [pagePath]="'/home'">
- * </dot-uve-palette-list>
+ *   [pagePath]="'/home'"
+ *   [variantId]="'1'" />
  * ```
  */
 @Component({
@@ -92,6 +94,7 @@ import { DotUVEPaletteContenttypeComponent } from '../dot-uve-palette-contenttyp
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class DotUvePaletteListComponent implements OnInit {
+    @ViewChild('menu') menu!: { toggle: (event: Event) => void };
     @ViewChild('favoritesPanel') favoritesPanel?: DotFavoriteSelectorComponent;
 
     $type = input.required<DotUVEPaletteListTypes>({ alias: 'listType' });
@@ -106,10 +109,11 @@ export class DotUvePaletteListComponent implements OnInit {
     readonly #destroyRef = inject(DestroyRef);
 
     readonly searchControl = new FormControl('', { nonNullable: true });
-
     protected readonly DotUVEPaletteListView = DotUVEPaletteListView;
     protected readonly LOADING_ROWS = LOADING_ROWS_MOCK;
 
+    protected readonly $skipNextSearch = signal(false);
+    protected readonly $contextMenuItems = signal<MenuItem[]>([]);
     protected readonly $contenttypes = this.#paletteListStore.contenttypes;
     protected readonly $contentlets = this.#paletteListStore.contentlets;
     protected readonly $pagination = this.#paletteListStore.pagination;
@@ -118,117 +122,111 @@ export class DotUvePaletteListComponent implements OnInit {
     protected readonly $isEmpty = this.#paletteListStore.$isEmpty;
     protected readonly $layoutMode = this.#paletteListStore.layoutMode;
     protected readonly $showListLayout = this.#paletteListStore.$showListLayout;
-    protected readonly $hideControls = computed(() => {
-        return this.$isEmpty() && !(this.searchControl.value.trim().length > 0);
+    protected readonly $isContentletsView = this.#paletteListStore.$isContentletsView;
+    protected readonly $isContentTypesView = this.#paletteListStore.$isContentTypesView;
+    protected readonly $isFavoritesList = this.#paletteListStore.$isFavoritesList;
+
+    /**
+     * Signal to determine if the search field has text.
+     * @returns True if the search field has non-whitespace content, false otherwise.
+     */
+    protected readonly $hasSearchText = toSignal(
+        this.searchControl.valueChanges.pipe(map((value) => value.trim().length > 0)),
+        { initialValue: false }
+    );
+
+    /**
+     * Computed signal to determine the start index for the pagination.
+     * @returns The start index for the pagination.
+     */
+    protected readonly $start = computed(() => {
+        const currentPage = this.$pagination().currentPage;
+        const perPage = this.$pagination().perPage;
+        return (currentPage - 1) * perPage;
     });
+
+    /**
+     * Computed signal to determine if the search and action button should be hidden.
+     * @returns True if the search and action button should be hidden, false otherwise.
+     */
+    protected readonly $hideControls = computed(() => {
+        return (this.$isEmpty() || this.$isLoading()) && !this.$hasSearchText();
+    });
+
+    /**
+     * Computed signal to determine the action button object.
+     * @returns The action button object.
+     */
+    protected readonly $actionButton = computed(() => {
+        if (this.$isFavoritesList()) {
+            return {
+                testId: 'add-favorites-button',
+                icon: 'pi pi-plus',
+                onClick: (event: Event) => this.favoritesPanel?.toggle(event)
+            };
+        }
+
+        return {
+            testId: 'sort-menu-button',
+            icon: 'pi pi-arrow-right-arrow-left',
+            onClick: (event: Event) => this.menu.toggle(event)
+        };
+    });
+
+    /**
+     * Computed signal to determine the menu items.
+     * @returns The menu items.
+     */
+    protected $menuItems = computed(() => {
+        const currentSort = this.#paletteListStore.$currentSort();
+        const viewMode = this.$layoutMode();
+        const onSortSelect = (sortOption: DotPaletteSortOption) => this.onSortSelect(sortOption);
+        const onViewSelect = (viewOption: DotPaletteViewMode) => this.onViewSelect(viewOption);
+
+        return buildPaletteMenuItems({ viewMode, currentSort, onSortSelect, onViewSelect });
+    });
+
+    /**
+     * Getter to determine the empty message object.
+     * @returns The empty message object.
+     */
     protected readonly $emptyState = computed(() => {
-        const searchTerm = this.searchControl.value.trim();
-        if (searchTerm.length > 0) {
+        if (this.$hasSearchText()) {
             return EMPTY_MESSAGE_SEARCH;
+        }
+
+        if (this.$isContentletsView()) {
+            return EMPTY_MESSAGE_CONTENTLETS;
         }
 
         return EMPTY_MESSAGES[this.$type()];
     });
 
-    protected readonly $contextMenuItems = signal<MenuItem[]>([]);
-
-    readonly $start = computed(
-        () => (this.$pagination().currentPage - 1) * this.$pagination().perPage
-    );
-    readonly $isFavoritesList = computed(() => this.$type() === DotUVEPaletteListTypes.FAVORITES);
-    readonly $showSortButton = computed(
-        () =>
-            this.$currentView() === DotUVEPaletteListView.CONTENT_TYPES && !this.$isFavoritesList()
-    );
-
-    protected $menuItems = computed(() => {
-        const viewMode = this.$layoutMode();
-        const currentSort = this.#paletteListStore.$currentSort();
-        return [
-            {
-                label: this.#dotMessageService.get('uve.palette.menu.sort.title'),
-                items: [
-                    {
-                        label: this.#dotMessageService.get('uve.palette.menu.sort.option.popular'),
-                        command: () => this.onSortSelect({ orderby: 'usage', direction: 'ASC' }),
-                        styleClass: getSortActiveClass(
-                            { orderby: 'usage', direction: 'ASC' },
-                            currentSort
-                        )
-                    },
-                    {
-                        label: this.#dotMessageService.get('uve.palette.menu.sort.option.a-to-z'),
-                        command: () => this.onSortSelect({ orderby: 'name', direction: 'ASC' }),
-                        styleClass: getSortActiveClass(
-                            { orderby: 'name', direction: 'ASC' },
-                            currentSort
-                        )
-                    },
-                    {
-                        label: this.#dotMessageService.get('uve.palette.menu.sort.option.z-to-a'),
-                        command: () => this.onSortSelect({ orderby: 'name', direction: 'DESC' }),
-                        styleClass: getSortActiveClass(
-                            { orderby: 'name', direction: 'DESC' },
-                            currentSort
-                        )
-                    }
-                ]
-            },
-            {
-                label: this.#dotMessageService.get('uve.palette.menu.view.title'),
-                items: [
-                    {
-                        label: this.#dotMessageService.get('uve.palette.menu.view.option.grid'),
-                        command: () => this.onViewSelect('grid'),
-                        styleClass: viewMode === 'grid' ? 'active-menu-item' : ''
-                    },
-                    {
-                        label: this.#dotMessageService.get('uve.palette.menu.view.option.list'),
-                        command: () => this.onViewSelect('list'),
-                        styleClass: viewMode === 'list' ? 'active-menu-item' : ''
-                    }
-                ]
-            }
-        ];
-    });
-
     constructor() {
         // React to input changes and fetch content types
         effect(() => {
-            const pagePathOrId = this.$pagePath();
-            const language = this.$languageId();
-            const variantId = this.$variantId();
-            const listType = this.$type();
+            const params = {
+                pagePathOrId: this.$pagePath(),
+                language: this.$languageId(),
+                variantId: this.$variantId(),
+                listType: this.$type()
+            };
 
             // Use untracked to prevent writes during effect
-            untracked(() => {
-                this.#paletteListStore.getContentTypes({
-                    pagePathOrId,
-                    language,
-                    variantId,
-                    listType
-                });
-            });
+            untracked(() => this.#paletteListStore.getContentTypes(params));
         });
     }
 
     ngOnInit() {
         // Set up debounced search with distinctUntilChanged to avoid duplicate calls
         this.searchControl.valueChanges
-            .pipe(debounceTime(300), distinctUntilChanged(), takeUntilDestroyed(this.#destroyRef))
-            .subscribe((searchTerm) => {
-                if (this.#paletteListStore.$isContentTypesView()) {
-                    this.#paletteListStore.getContentTypes({
-                        filter: searchTerm,
-                        page: 1 // Reset to first page on search
-                    });
-                } else {
-                    this.#paletteListStore.getContentlets({
-                        filter: searchTerm,
-                        page: 1
-                    });
-                }
-            });
+            .pipe(
+                debounceTime(300),
+                distinctUntilChanged(),
+                filter(() => this.#shouldRunSearch()),
+                takeUntilDestroyed(this.#destroyRef)
+            )
+            .subscribe((filter) => this.#loadItems({ filter, page: 1 }));
     }
 
     /**
@@ -240,11 +238,7 @@ export class DotUvePaletteListComponent implements OnInit {
     protected onPageChange(event: PaginatorState) {
         const page = (event.page ?? 0) + 1; // PrimeNG uses 0-based pages
 
-        if (this.#paletteListStore.$isContentTypesView()) {
-            this.#paletteListStore.getContentTypes({ page });
-        } else {
-            this.#paletteListStore.getContentlets({ page });
-        }
+        this.#loadItems({ page });
     }
 
     /**
@@ -277,13 +271,13 @@ export class DotUvePaletteListComponent implements OnInit {
      *
      * @param contentTypeName - The name of the content type to drill into
      */
-    protected onSelectContentType(contentTypeName: string) {
+    protected onSelectContentType(selectedContentType: string) {
         this.#paletteListStore.getContentlets({
-            selectedContentType: contentTypeName,
+            selectedContentType,
             filter: '',
             page: 1
         });
-        this.searchControl.setValue('', { emitEvent: false });
+        this.#resetSearch();
     }
 
     /**
@@ -296,7 +290,7 @@ export class DotUvePaletteListComponent implements OnInit {
             filter: '',
             page: 1
         });
-        this.searchControl.setValue('', { emitEvent: false });
+        this.#resetSearch();
     }
 
     protected onContextMenu(contentType: DotCMSContentType) {
@@ -305,47 +299,9 @@ export class DotUvePaletteListComponent implements OnInit {
             ? 'uve.palette.menu.favorite.option.remove'
             : 'uve.palette.menu.favorite.option.add';
         const command = isFavorite
-            ? () => this.removeFavoriteItems(contentType)
-            : () => this.addFavoriteItems(contentType);
+            ? () => this.#removeFavorite(contentType)
+            : () => this.#addFavorite(contentType);
         this.$contextMenuItems.set([{ label: this.#dotMessageService.get(label), command }]);
-    }
-
-    /**
-     * Remove a content type from favorites.
-     * @param contentType - The content type to remove.
-     */
-    private removeFavoriteItems(contentType: DotCMSContentType) {
-        const contenttypes = this.#dotFavoriteContentTypeService.remove(contentType.id);
-
-        if (this.$type() === DotUVEPaletteListTypes.FAVORITES) {
-            this.#paletteListStore.setContentTypesFromFavorite(contenttypes);
-        }
-
-        this.#messageService.add({
-            severity: 'success',
-            summary: this.#dotMessageService.get('uve.palette.favorite.remove.success.summary'),
-            detail: this.#dotMessageService.get('uve.palette.favorite.remove.success.detail'),
-            life: 3000
-        });
-    }
-
-    /**
-     * Add a content type to favorites.
-     * @param contentType - The content type to add.
-     */
-    private addFavoriteItems(contentType: DotCMSContentType) {
-        const contenttypes = this.#dotFavoriteContentTypeService.add(contentType);
-
-        if (this.$type() === DotUVEPaletteListTypes.FAVORITES) {
-            this.#paletteListStore.setContentTypesFromFavorite(contenttypes);
-        }
-
-        this.#messageService.add({
-            severity: 'success',
-            summary: this.#dotMessageService.get('uve.palette.favorite.add.success.summary'),
-            detail: this.#dotMessageService.get('uve.palette.favorite.add.success.detail'),
-            life: 3000
-        });
     }
 
     /**
@@ -362,5 +318,75 @@ export class DotUvePaletteListComponent implements OnInit {
         }
 
         this.favoritesPanel?.toggle(event);
+    }
+
+    /**
+     * Remove a content type from favorites.
+     * @param contentType - The content type to remove.
+     */
+    #removeFavorite(contentType: DotCMSContentType) {
+        this.#paletteListStore.removeFavorite(contentType.id);
+        this.#messageService.add({
+            severity: 'success',
+            summary: this.#dotMessageService.get('uve.palette.favorite.remove.success.summary'),
+            detail: this.#dotMessageService.get('uve.palette.favorite.remove.success.detail'),
+            life: 3000
+        });
+    }
+
+    /**
+     * Add a content type to favorites.
+     * @param contentType - The content type to add.
+     */
+    #addFavorite(contentType: DotCMSContentType) {
+        this.#paletteListStore.addFavorite(contentType);
+        this.#messageService.add({
+            severity: 'success',
+            summary: this.#dotMessageService.get('uve.palette.favorite.add.success.summary'),
+            detail: this.#dotMessageService.get('uve.palette.favorite.add.success.detail'),
+            life: 3000
+        });
+    }
+
+    /**
+     * Clears the search input without triggering another request.
+     * Keeps debounced listeners quiet when switching views manually.
+     */
+    #resetSearch() {
+        if (!this.$hasSearchText()) {
+            // Search is already empty, nothing to do
+            return;
+        }
+
+        // Search has text, clear it and skip the debounced search trigger
+        this.$skipNextSearch.set(true);
+        this.searchControl.setValue('');
+    }
+
+    /**
+     * Dispatches the appropriate store fetch based on the current view.
+     * Accepts any partial search params so pagination and filtering can reuse it.
+     *
+     * @param params - Partial palette search params (filter/page/order/etc.)
+     */
+    #loadItems(params: Partial<DotPaletteSearchParams>) {
+        if (this.$isContentTypesView()) {
+            this.#paletteListStore.getContentTypes(params);
+            return;
+        }
+
+        this.#paletteListStore.getContentlets(params);
+    }
+
+    /**
+     * Determines if the search should be run.
+     * @returns True if the search should be run, false otherwise.
+     */
+    #shouldRunSearch() {
+        if (this.$skipNextSearch()) {
+            this.$skipNextSearch.set(false);
+            return false;
+        }
+        return true;
     }
 }
