@@ -10,7 +10,7 @@ describe('ContentSearchService', () => {
     });
 
     describe('search', () => {
-        it('should search content successfully', async () => {
+        it('should search content successfully with published content', async () => {
             const mockResponse = {
                 json: jest.fn().mockResolvedValue({
                     entity: {
@@ -19,11 +19,10 @@ describe('ContentSearchService', () => {
                         jsonObjectView: {
                             contentlets: [
                                 {
-                                    id: '1',
                                     title: 'Test Result',
                                     hostName: 'localhost',
-                                    modDate: '20240101',
-                                    publishDate: '20240101',
+                                    modDate: '1764284834965',
+                                    publishDate: '1764284834994',
                                     baseType: 'CONTENT',
                                     inode: 'inode1',
                                     archived: false,
@@ -39,10 +38,10 @@ describe('ContentSearchService', () => {
                                     publishUserName: 'admin',
                                     publishUser: 'admin',
                                     languageId: 1,
-                                    creationDate: '20240101',
+                                    creationDate: '1764284749322',
                                     shortyId: 'shorty1',
                                     url: '/test',
-                                    titleImage: '',
+                                    titleImage: 'TITLE_IMAGE_NOT_FOUND',
                                     modUserName: 'admin',
                                     hasLiveVersion: true,
                                     folder: '/test',
@@ -83,6 +82,223 @@ describe('ContentSearchService', () => {
             );
             expect(result.entity?.jsonObjectView?.contentlets).toHaveLength(1);
             expect(result.entity?.jsonObjectView?.contentlets?.[0].title).toBe('Test Result');
+            // Verify dates are converted to numbers
+            expect(typeof result.entity?.jsonObjectView?.contentlets?.[0].modDate).toBe('number');
+            expect(typeof result.entity?.jsonObjectView?.contentlets?.[0].creationDate).toBe(
+                'number'
+            );
+            expect(typeof result.entity?.jsonObjectView?.contentlets?.[0].publishDate).toBe(
+                'number'
+            );
+        });
+
+        it('should search content successfully with unpublished content', async () => {
+            const mockResponse = {
+                json: jest.fn().mockResolvedValue({
+                    entity: {
+                        contentTook: 1,
+                        resultsSize: 1,
+                        jsonObjectView: {
+                            contentlets: [
+                                {
+                                    title: 'Unpublished Result',
+                                    hostName: 'localhost',
+                                    modDate: '1764284834965',
+                                    baseType: 'CONTENT',
+                                    inode: 'inode2',
+                                    archived: false,
+                                    host: 'host1',
+                                    ownerUserName: 'admin',
+                                    working: true,
+                                    locked: false,
+                                    stInode: 'stinode1',
+                                    contentType: 'Blog',
+                                    live: false,
+                                    owner: 'admin',
+                                    identifier: 'id2',
+                                    languageId: 1,
+                                    creationDate: '1764284749322',
+                                    shortyId: 'shorty2',
+                                    url: '/test2',
+                                    titleImage: 'TITLE_IMAGE_NOT_FOUND',
+                                    modUserName: 'admin',
+                                    hasLiveVersion: false,
+                                    folder: '/test',
+                                    hasTitleImage: false,
+                                    sortOrder: 0,
+                                    modUser: 'admin',
+                                    __icon__: 'icon',
+                                    contentTypeIcon: 'icon',
+                                    variant: 'default'
+                                }
+                            ]
+                        },
+                        queryTook: 1
+                    },
+                    errors: [],
+                    messages: [],
+                    i18nMessagesMap: {},
+                    permissions: []
+                })
+            };
+            mockFetch.mockResolvedValue(mockResponse);
+
+            const params = {
+                query: '+title:Unpublished',
+                limit: 1,
+                languageId: 1,
+                depth: 1,
+                allCategoriesInfo: false
+            };
+            const result = await service.search(params);
+
+            expect(result.entity?.jsonObjectView?.contentlets).toHaveLength(1);
+            expect(result.entity?.jsonObjectView?.contentlets?.[0].title).toBe(
+                'Unpublished Result'
+            );
+            // Verify publishDate is null/undefined for unpublished content
+            expect(result.entity?.jsonObjectView?.contentlets?.[0].publishDate).toBeUndefined();
+            // Verify dates are converted to numbers
+            expect(typeof result.entity?.jsonObjectView?.contentlets?.[0].modDate).toBe('number');
+            expect(typeof result.entity?.jsonObjectView?.contentlets?.[0].creationDate).toBe(
+                'number'
+            );
+        });
+
+        it('should handle numeric date values', async () => {
+            const mockResponse = {
+                json: jest.fn().mockResolvedValue({
+                    entity: {
+                        contentTook: 1,
+                        resultsSize: 1,
+                        jsonObjectView: {
+                            contentlets: [
+                                {
+                                    title: 'Numeric Dates',
+                                    hostName: 'localhost',
+                                    modDate: 1764284834965,
+                                    publishDate: 1764284834994,
+                                    baseType: 'CONTENT',
+                                    inode: 'inode3',
+                                    archived: false,
+                                    host: 'host1',
+                                    ownerUserName: 'admin',
+                                    working: true,
+                                    locked: false,
+                                    stInode: 'stinode1',
+                                    contentType: 'Blog',
+                                    live: true,
+                                    owner: 'admin',
+                                    identifier: 'id3',
+                                    publishUserName: 'admin',
+                                    publishUser: 'admin',
+                                    languageId: 1,
+                                    creationDate: 1764284749322,
+                                    shortyId: 'shorty3',
+                                    url: '/test3',
+                                    titleImage: 'TITLE_IMAGE_NOT_FOUND',
+                                    modUserName: 'admin',
+                                    hasLiveVersion: true,
+                                    folder: '/test',
+                                    hasTitleImage: false,
+                                    sortOrder: 0,
+                                    modUser: 'admin',
+                                    __icon__: 'icon',
+                                    contentTypeIcon: 'icon',
+                                    variant: 'default'
+                                }
+                            ]
+                        },
+                        queryTook: 1
+                    },
+                    errors: [],
+                    messages: [],
+                    i18nMessagesMap: {},
+                    permissions: []
+                })
+            };
+            mockFetch.mockResolvedValue(mockResponse);
+
+            const params = {
+                query: '+title:Numeric',
+                limit: 1,
+                languageId: 1,
+                depth: 1,
+                allCategoriesInfo: false
+            };
+            const result = await service.search(params);
+
+            expect(result.entity?.jsonObjectView?.contentlets).toHaveLength(1);
+            expect(result.entity?.jsonObjectView?.contentlets?.[0].modDate).toBe(1764284834965);
+            expect(result.entity?.jsonObjectView?.contentlets?.[0].creationDate).toBe(
+                1764284749322
+            );
+            expect(result.entity?.jsonObjectView?.contentlets?.[0].publishDate).toBe(1764284834994);
+        });
+
+        it('should handle null publishDate', async () => {
+            const mockResponse = {
+                json: jest.fn().mockResolvedValue({
+                    entity: {
+                        contentTook: 1,
+                        resultsSize: 1,
+                        jsonObjectView: {
+                            contentlets: [
+                                {
+                                    title: 'Null Publish Date',
+                                    hostName: 'localhost',
+                                    modDate: '1764284834965',
+                                    publishDate: null,
+                                    baseType: 'CONTENT',
+                                    inode: 'inode4',
+                                    archived: false,
+                                    host: 'host1',
+                                    ownerUserName: 'admin',
+                                    working: true,
+                                    locked: false,
+                                    stInode: 'stinode1',
+                                    contentType: 'Blog',
+                                    live: false,
+                                    owner: 'admin',
+                                    identifier: 'id4',
+                                    languageId: 1,
+                                    creationDate: '1764284749322',
+                                    shortyId: 'shorty4',
+                                    url: '/test4',
+                                    titleImage: 'TITLE_IMAGE_NOT_FOUND',
+                                    modUserName: 'admin',
+                                    hasLiveVersion: false,
+                                    folder: '/test',
+                                    hasTitleImage: false,
+                                    sortOrder: 0,
+                                    modUser: 'admin',
+                                    __icon__: 'icon',
+                                    contentTypeIcon: 'icon',
+                                    variant: 'default'
+                                }
+                            ]
+                        },
+                        queryTook: 1
+                    },
+                    errors: [],
+                    messages: [],
+                    i18nMessagesMap: {},
+                    permissions: []
+                })
+            };
+            mockFetch.mockResolvedValue(mockResponse);
+
+            const params = {
+                query: '+title:Null',
+                limit: 1,
+                languageId: 1,
+                depth: 1,
+                allCategoriesInfo: false
+            };
+            const result = await service.search(params);
+
+            expect(result.entity?.jsonObjectView?.contentlets).toHaveLength(1);
+            expect(result.entity?.jsonObjectView?.contentlets?.[0].publishDate).toBeNull();
         });
 
         it('should handle invalid search parameters', async () => {
