@@ -4,8 +4,23 @@ import { z } from 'zod';
 export const ContentletSchema = z
     .object({
         hostName: z.string(),
-        modDate: z.string(),
-        publishDate: z.string(),
+        modDate: z.preprocess((val) => {
+            if (typeof val === 'number') return val;
+            if (typeof val === 'string' && val) {
+                const num = Number(val);
+                return isNaN(num) ? 0 : num;
+            }
+            return 0;
+        }, z.number()),
+        publishDate: z.preprocess((val) => {
+            if (val === null || val === undefined) return val;
+            if (typeof val === 'number') return val;
+            if (typeof val === 'string' && val) {
+                const num = Number(val);
+                return isNaN(num) ? null : num;
+            }
+            return null;
+        }, z.number().nullish()),
         title: z.string(),
         baseType: z.string(),
         inode: z.string(),
@@ -19,10 +34,17 @@ export const ContentletSchema = z
         live: z.boolean(),
         owner: z.string(),
         identifier: z.string(),
-        publishUserName: z.string(),
-        publishUser: z.string(),
+        publishUserName: z.string().nullish(),
+        publishUser: z.string().nullish(),
         languageId: z.number(),
-        creationDate: z.string(),
+        creationDate: z.preprocess((val) => {
+            if (typeof val === 'number') return val;
+            if (typeof val === 'string' && val) {
+                const num = Number(val);
+                return isNaN(num) ? 0 : num;
+            }
+            return 0;
+        }, z.number()),
         shortyId: z.string(),
         url: z.string(),
         titleImage: z.string(),
