@@ -17,6 +17,8 @@ import { filter, map } from 'rxjs/operators';
 
 import { MenuItemEntity } from '@dotcms/dotcms-models';
 
+import { processSpecialRoute } from './breadcrumb.utils';
+
 /**
  * State interface for the Breadcrumb feature.
  * Contains the breadcrumb navigation items.
@@ -101,7 +103,15 @@ export function withBreadcrumbs(menuItems: Signal<MenuItemEntity[]>) {
         }),
         withMethods((store) => {
             const setBreadcrumbs = (breadcrumbs: MenuItem[]) => {
-                patchState(store, { breadcrumbs });
+                patchState(store, {
+                    breadcrumbs: [
+                        {
+                            label: 'Home',
+                            disabled: true
+                        },
+                        ...breadcrumbs
+                    ]
+                });
             };
 
             const appendCrumb = (crumb: MenuItem) => {
@@ -170,94 +180,65 @@ export function withBreadcrumbs(menuItems: Signal<MenuItemEntity[]>) {
                     return;
                 }
 
-                const newUrl = `/dotAdmin/#${url}`;
                 const breadcrumbs = store.breadcrumbs();
+                const [urlPath, queryString] = url.split('?');
+                const shortMenuId = new URLSearchParams(queryString || '').get('mId');
 
-                const existingIndex = breadcrumbs.findIndex((crumb) => {
-                    return crumb.url === newUrl;
-                });
-
-                if (existingIndex > -1) {
-                    truncateBreadcrumbs(existingIndex);
-                } else {
-                    const [urlPath, queryString] = url.split('?');
-                    const shortMenuId = new URLSearchParams(queryString || '').get('mId');
-
+                const item = menu.find((item) => {
+                    const pathMatches = item.menuLink === urlPath;
                     const hasQueryParams = queryString && queryString.length > 0;
 
-                    const item = menu.find((item) => {
-                        const pathMatches = item.menuLink === urlPath;
+                    // If we have query params but no mId, it's likely an old bookmark - don't match
+                    if (hasQueryParams && !shortMenuId) {
+                        return false;
+                    }
 
-                        // If we have query params but no mId, it's likely an old bookmark - don't match
-                        if (hasQueryParams && !shortMenuId) {
-                            return false;
+                    // If we have mId, validate both path and parent match
+                    if (shortMenuId) {
+                        return pathMatches && item.parentMenuId.startsWith(shortMenuId);
+                    }
+
+                    // Default: no query params, no mId - match by path only
+                    return pathMatches;
+                });
+
+                if (item) {
+                    const newUrl = `/dotAdmin/#${url}`;
+                    setBreadcrumbs([
+                        {
+                            label: item.parentMenuLabel,
+                            disabled: true
+                        },
+                        {
+                            label: item.label,
+                            target: '_self',
+                            url: newUrl
                         }
-
-                        // If we have mId, validate both path and parent match
-                        if (shortMenuId) {
-                            return pathMatches && item.parentMenuId.startsWith(shortMenuId);
-                        }
-
-                        // Default: no query params, no mId - match by path only
-                        return pathMatches;
+                    ]);
+                } else {
+                    const result = processSpecialRoute({
+                        url,
+                        menu,
+                        breadcrumbs
                     });
 
-                    if (item) {
-                        setBreadcrumbs([
-                            {
-                                label: 'Home',
-                                disabled: true
-                            },
-                            {
-                                label: item.parentMenuLabel,
-                                disabled: true
-                            },
-                            {
-                                label: item.label,
-                                target: '_self',
-                                url: newUrl
+                    if (result) {
+                        switch (result.type) {
+                            case 'set':
+                                setBreadcrumbs(result.breadcrumbs);
+
+                                break;
+                            case 'append': {
+                                const crumb = result.breadcrumbs[result.breadcrumbs.length - 1];
+                                appendCrumb(crumb);
+
+                                break;
                             }
-                        ]);
-                    } else {
-                        // Handle special case: /templates/edit/:id
-                        const templatesEditRegex = /^\/templates\/edit\/[a-zA-Z0-9-]+$/;
-
-                        if (templatesEditRegex.test(url)) {
-                            const templatesItem = menu.find(
-                                (item) => item.menuLink === '/templates'
-                            );
-
-                            if (templatesItem) {
-                                // Only build base breadcrumb if it doesn't exist yet
-                                const hasTemplatesBreadcrumb = breadcrumbs.some(
-                                    (crumb) => crumb.url === '/dotAdmin/#/templates'
-                                );
-
-                                if (!hasTemplatesBreadcrumb) {
-                                    setBreadcrumbs([
-                                        {
-                                            label: 'Home',
-                                            disabled: true
-                                        },
-                                        {
-                                            label: templatesItem.parentMenuLabel,
-                                            disabled: true
-                                        },
-                                        {
-                                            label: templatesItem.label,
-                                            target: '_self',
-                                            url: '/dotAdmin/#/templates'
-                                        }
-                                    ]);
+                            case 'truncate':
+                                if (result.index && result.index > -1) {
+                                    truncateBreadcrumbs(result.index);
                                 }
-                            }
-                        } else if (url.includes('/content?filter=')) {
-                            const filter = url.split('/content?filter=')[1];
-                            addNewBreadcrumb({
-                                label: filter,
-                                target: '_self',
-                                url: newUrl
-                            });
+                                break;
                         }
                     }
                 }
