@@ -1,3 +1,5 @@
+import { MenuItem } from 'primeng/api';
+
 import {
     DEFAULT_VARIANT_ID,
     DotCMSContentlet,
@@ -9,6 +11,7 @@ import {
     DEFAULT_PER_PAGE,
     DotPaletteSortOption,
     DotPaletteListStatus,
+    DotPaletteViewMode,
     DotUVEPaletteListTypes
 } from '../models';
 
@@ -51,7 +54,8 @@ export const EMPTY_CONTENTTYPE_RESPONSE = {
  */
 export const EMPTY_CONTENTLET_RESPONSE = {
     contentlets: [] as DotCMSContentlet[],
-    pagination: EMPTY_PAGINATION
+    pagination: EMPTY_PAGINATION,
+    status: DotPaletteListStatus.EMPTY
 };
 
 export const DEFAULT_SORT_OPTIONS: DotPaletteSortOption = {
@@ -86,6 +90,65 @@ export function getSortActiveClass(
     const isActive = sameOrderby && sameDirection;
 
     return isActive ? 'active-menu-item' : '';
+}
+
+export function buildPaletteMenuItems({
+    viewMode,
+    currentSort,
+    onSortSelect,
+    onViewSelect
+}: {
+    viewMode: DotPaletteViewMode;
+    currentSort: DotPaletteSortOption;
+    onSortSelect: (sortOption: DotPaletteSortOption) => void;
+    onViewSelect: (viewMode: DotPaletteViewMode) => void;
+}): MenuItem[] {
+    return [
+        {
+            label: 'uve.palette.menu.sort.title',
+            items: [
+                {
+                    label: 'uve.palette.menu.sort.option.popular',
+                    command: () => onSortSelect({ orderby: 'usage', direction: 'ASC' }),
+                    styleClass: getSortActiveClass(
+                        { orderby: 'usage', direction: 'ASC' },
+                        currentSort
+                    )
+                },
+                {
+                    label: 'uve.palette.menu.sort.option.a-to-z',
+                    command: () => onSortSelect({ orderby: 'name', direction: 'ASC' }),
+                    styleClass: getSortActiveClass(
+                        { orderby: 'name', direction: 'ASC' },
+                        currentSort
+                    )
+                },
+                {
+                    label: 'uve.palette.menu.sort.option.z-to-a',
+                    command: () => onSortSelect({ orderby: 'name', direction: 'DESC' }),
+                    styleClass: getSortActiveClass(
+                        { orderby: 'name', direction: 'DESC' },
+                        currentSort
+                    )
+                }
+            ]
+        },
+        {
+            label: 'uve.palette.menu.view.title',
+            items: [
+                {
+                    label: 'uve.palette.menu.view.option.grid',
+                    command: () => onViewSelect('grid'),
+                    styleClass: viewMode === 'grid' ? 'active-menu-item' : ''
+                },
+                {
+                    label: 'uve.palette.menu.view.option.list',
+                    command: () => onViewSelect('list'),
+                    styleClass: viewMode === 'list' ? 'active-menu-item' : ''
+                }
+            ]
+        }
+    ];
 }
 
 /**
@@ -125,7 +188,7 @@ export function getPaletteState(
  *
  * @example
  * ```typescript
- * const result = filterAndBuildFavoriteResponse({
+ * const result = buildPaletteFavorite({
  *   contentTypes: allContentTypes,
  *   filter: 'blog',
  *   page: 2
@@ -136,7 +199,7 @@ export function getPaletteState(
  * // }
  * ```
  */
-export function filterAndBuildFavoriteResponse({
+export function buildPaletteFavorite({
     contentTypes,
     filter = '',
     page = 1
@@ -147,6 +210,7 @@ export function filterAndBuildFavoriteResponse({
 }): {
     contenttypes: DotCMSContentType[];
     pagination: { currentPage: number; perPage: number; totalEntries: number };
+    status: DotPaletteListStatus;
 } {
     // Filter and sort all content types
     const filteredContentTypes = contentTypes.filter(
@@ -166,7 +230,7 @@ export function filterAndBuildFavoriteResponse({
         totalEntries
     };
 
-    return { contenttypes, pagination };
+    return { contenttypes, pagination, status: getPaletteState(contenttypes) };
 }
 
 /**
@@ -189,12 +253,13 @@ export function filterAndBuildFavoriteResponse({
  * // }
  * ```
  */
-export function buildContentletsResponse(
+export function buildPaletteContent(
     response: ESContent,
     offset: number
 ): {
     contentlets: DotCMSContentlet[];
     pagination: { currentPage: number; perPage: number; totalEntries: number };
+    status: DotPaletteListStatus;
 } {
     const contentlets = response.jsonObjectView.contentlets;
     const totalEntries = response.resultsSize;
@@ -202,7 +267,8 @@ export function buildContentletsResponse(
 
     return {
         contentlets,
-        pagination: { currentPage, perPage: contentlets.length, totalEntries }
+        pagination: { currentPage, perPage: contentlets.length, totalEntries },
+        status: getPaletteState(contentlets)
     };
 }
 
@@ -296,6 +362,12 @@ export const EMPTY_MESSAGE_SEARCH = {
     icon: 'pi pi-search',
     title: 'uve.palette.empty.search.state.title',
     message: 'uve.palette.empty.search.state.message'
+};
+
+export const EMPTY_MESSAGE_CONTENTLETS = {
+    icon: 'pi pi-folder-open',
+    title: 'uve.palette.empty.state.contentlets.title',
+    message: 'uve.palette.empty.state.contentlets.message'
 };
 
 /**
