@@ -1,12 +1,15 @@
 import { describe } from '@jest/globals';
 import { createServiceFactory, mockProvider, SpectatorService } from '@ngneat/spectator/jest';
-import { signalStore, withState } from '@ngrx/signals';
+import { patchState, signalStore, withState } from '@ngrx/signals';
 
 import { ActivatedRoute, Router } from '@angular/router';
+
+import { UVE_MODE } from '@dotcms/types';
 
 import { withClient } from './withClient';
 
 import { DotPageApiParams } from '../../../services/dot-page-api.service';
+import { PERSONA_KEY } from '../../../shared/consts';
 import { UVE_STATUS } from '../../../shared/enums';
 import { UVEState } from '../../models';
 
@@ -27,7 +30,11 @@ const initialState: UVEState = {
     isClientReady: false
 };
 
-export const uveStoreMock = signalStore(withState<UVEState>(initialState), withClient());
+export const uveStoreMock = signalStore(
+    { protectedState: false },
+    withState<UVEState>(initialState),
+    withClient()
+);
 
 describe('UVEStore', () => {
     let spectator: SpectatorService<InstanceType<typeof uveStoreMock>>;
@@ -83,6 +90,120 @@ describe('UVEStore', () => {
             store.resetClientConfiguration();
 
             expect(store.graphql()).toEqual(null);
+        });
+    });
+
+    describe('$graphqlWithParams', () => {
+        it('should return null when graphql is null', () => {
+            expect(store.$graphqlWithParams()).toBeNull();
+        });
+
+        it('should return null when graphql is not set', () => {
+            store.resetClientConfiguration();
+            expect(store.$graphqlWithParams()).toBeNull();
+        });
+
+        it('should merge graphql variables with page params', () => {
+            const graphql = {
+                query: 'test query',
+                variables: {
+                    depth: '1'
+                }
+            };
+
+            const pageParams: DotPageApiParams = {
+                url: '/test-url',
+                mode: UVE_MODE.EDIT,
+                language_id: '1',
+                variantName: 'DEFAULT',
+                [PERSONA_KEY]: 'persona-id-123'
+            };
+
+            patchState(store, { pageParams });
+            store.setCustomGraphQL(graphql, false);
+
+            const result = store.$graphqlWithParams();
+
+            expect(result).toEqual({
+                query: 'test query',
+                variables: {
+                    depth: '1',
+                    url: '/test-url',
+                    mode: UVE_MODE.EDIT,
+                    languageId: '1',
+                    variantName: 'DEFAULT',
+                    personaId: 'persona-id-123'
+                }
+            });
+        });
+
+        it('should preserve existing graphql variables when merging with page params', () => {
+            const graphql = {
+                query: 'test query',
+                variables: {
+                    depth: '2',
+                    customVar: 'custom-value'
+                }
+            };
+
+            const pageParams: DotPageApiParams = {
+                url: '/another-url',
+                mode: UVE_MODE.PREVIEW,
+                language_id: '2',
+                variantName: 'VARIANT_A',
+                [PERSONA_KEY]: 'persona-id-456'
+            };
+
+            patchState(store, { pageParams });
+            store.setCustomGraphQL(graphql, false);
+
+            const result = store.$graphqlWithParams();
+
+            expect(result).toEqual({
+                query: 'test query',
+                variables: {
+                    depth: '2',
+                    customVar: 'custom-value',
+                    url: '/another-url',
+                    mode: UVE_MODE.PREVIEW,
+                    languageId: '2',
+                    variantName: 'VARIANT_A',
+                    personaId: 'persona-id-456'
+                }
+            });
+        });
+
+        it('should handle missing optional page params', () => {
+            const graphql = {
+                query: 'test query',
+                variables: {
+                    depth: '1'
+                }
+            };
+
+            const pageParams: DotPageApiParams = {
+                url: '/test-url',
+                language_id: '1',
+                [PERSONA_KEY]: 'persona-id-123'
+                // mode and variantName are missing
+            };
+
+            patchState(store, { pageParams });
+            store.setCustomGraphQL(graphql, false);
+
+            const result = store.$graphqlWithParams();
+
+            expect(result).toEqual({
+                query: 'test query',
+                variables: {
+                    depth: '1',
+                    url: '/test-url',
+                    mode: undefined,
+                    languageId: '1',
+                    variantName: undefined,
+                    personaId: 'persona-id-123'
+                }
+            });
         });
     });
 });
