@@ -7,7 +7,6 @@ import com.dotcms.variant.VariantAPI;
 import com.dotmarketing.business.DotStateException;
 import com.dotmarketing.portlets.contentlet.model.Contentlet;
 import com.dotmarketing.portlets.contentlet.transform.DotTransformerBuilder;
-import com.dotmarketing.portlets.contentlet.util.ContentletUtil;
 import com.dotmarketing.util.Logger;
 import com.dotmarketing.util.UtilMethods;
 import com.dotmarketing.util.json.JSONObject;
@@ -106,16 +105,26 @@ public class ContentMapDataFetcher implements DataFetcher<Object> {
             final String mapKey = entry.getKey();
             final Object rawValue = entry.getValue();
 
+            // Searches if the map contains any key ending with "_raw".
             if (mapKey.endsWith("_raw") && rawValue instanceof String) {
+                // Creates a baseKey variable deleting the "_raw" string at the end of the mapKey. i.e. blogContent_raw = blogContent
                 final String baseKey = mapKey.substring(0, mapKey.length() - 4);
 
+                // Checks if the baseKey exist in the map
                 if (hydratedMap.containsKey(baseKey)) {
+                    // Checks if the baseValue (hydrated) is not empty. If the baseValue is empty or null, we parse the rawValue instead.
+                    Object hydratedValue = hydratedMap.get(baseKey);
+                    final String baseValue = (hydratedValue == null || hydratedValue.toString().trim().isEmpty())
+                            ? rawValue.toString()
+                            : hydratedValue.toString();
+
+                    // Parse the baseValue as JSON
                     try {
                         @SuppressWarnings("unchecked")
-                        Map<String, Object> parsed = objectMapper.readValue((String) rawValue, Map.class);
+                        Map<String, Object> parsed = objectMapper.readValue(baseValue, Map.class);
                         hydratedMap.put(baseKey, parsed);
                     } catch (Exception e) {
-                        Logger.warn(this, () -> "Error parsing JSON for '" + mapKey + "': " + e.getMessage());
+                        Logger.warn(this, () -> "Error parsing JSON for '" + baseKey + "': " + e.getMessage());
                     }
                 }
             }
@@ -151,7 +160,6 @@ public class ContentMapDataFetcher implements DataFetcher<Object> {
      * </ul>
      *
      * @param request    the current HTTP request, used to extract the variant key if provided
-     * @param user       the user requesting the content, used for permissions and rendering
      * @param contentlet the contentlet to hydrate
      * @param render     whether renderable fields should be velocity-rendered
      * @return a hydrated content map including field values (rendered if requested)
