@@ -1,3 +1,3 @@
 (function(){
-var c=(r=>(r.EDIT="EDIT_MODE",r.PREVIEW="PREVIEW_MODE",r.LIVE="LIVE",r.UNKNOWN="UNKNOWN",r))(c||{});function f(t){return t.map(e=>{let n=e.getBoundingClientRect(),o=Array.from(e.querySelectorAll('[data-dot-object="contentlet"]'));return{x:n.x,y:n.y,width:n.width,height:n.height,payload:JSON.stringify({container:T(e)}),contentlets:F(n,o)}})}function F(t,e){return e.map(n=>{let o=n.getBoundingClientRect();return{x:0,y:o.y-t.y,width:o.width,height:o.height,payload:JSON.stringify({container:n.dataset?.dotContainer?JSON.parse(n.dataset?.dotContainer):g(n),contentlet:{identifier:n.dataset?.dotIdentifier,title:n.dataset?.dotTitle,inode:n.dataset?.dotInode,contentType:n.dataset?.dotType}})}})}function T(t){return{acceptTypes:t.dataset?.dotAcceptTypes||"",identifier:t.dataset?.dotIdentifier||"",maxContentlets:t.dataset?.maxContentlets||"",uuid:t.dataset?.dotUuid||""}}function g(t){let e=t.closest('[data-dot-object="container"]');return e?T(e):(console.warn("No container found for the contentlet"),null)}function u(t){if(!t)return null;let e=t.querySelector('[data-dot-object="empty-content"]');return t?.dataset?.dotObject==="contentlet"||t?.dataset?.dotObject==="container"&&e||t?.dataset?.dotObject==="container"&&t.children.length===0?t:u(t?.parentElement)}function M(t){let e=t.querySelectorAll('[data-dot-object="vtl-file"]');return e.length?Array.from(e).map(n=>({inode:n.dataset?.dotInode,name:n.dataset?.dotUrl})):null}function h(){let t=document.documentElement.scrollHeight,e=window.innerHeight;return window.scrollY+e>=t}function x(t){let e=n=>{n.data.name==="uve-set-page-data"&&t(n.data.payload)};return window.addEventListener("message",e),{unsubscribe:()=>{window.removeEventListener("message",e)},event:"changes"}}function v(t){let e=n=>{n.data.name==="uve-reload-page"&&t()};return window.addEventListener("message",e),{unsubscribe:()=>{window.removeEventListener("message",e)},event:"page-reload"}}function N(t){let e=n=>{if(n.data.name==="uve-request-bounds"){let o=Array.from(document.querySelectorAll('[data-dot-object="container"]')),r=f(o);t(r)}};return window.addEventListener("message",e),{unsubscribe:()=>{window.removeEventListener("message",e)},event:"request-bounds"}}function O(t){let e=n=>{if(n.data.name==="uve-scroll-inside-iframe"){let o=n.data.direction;t(o)}};return window.addEventListener("message",e),{unsubscribe:()=>{window.removeEventListener("message",e)},event:"iframe-scroll"}}function L(t){let e=n=>{let o=u(n.target);if(!o)return;let{x:r,y:s,width:d,height:l}=o.getBoundingClientRect(),p=o.dataset?.dotObject==="container",H={identifier:"TEMP_EMPTY_CONTENTLET",title:"TEMP_EMPTY_CONTENTLET",contentType:"TEMP_EMPTY_CONTENTLET_TYPE",inode:"TEMPY_EMPTY_CONTENTLET_INODE",widgetTitle:"TEMP_EMPTY_CONTENTLET",baseType:"TEMP_EMPTY_CONTENTLET",onNumberOfPages:1},B={identifier:o.dataset?.dotIdentifier,title:o.dataset?.dotTitle,inode:o.dataset?.dotInode,contentType:o.dataset?.dotType,baseType:o.dataset?.dotBasetype,widgetTitle:o.dataset?.dotWidgetTitle,onNumberOfPages:o.dataset?.dotOnNumberOfPages},q=M(o),G={container:o.dataset?.dotContainer?JSON.parse(o.dataset?.dotContainer):g(o),contentlet:p?H:B,vtlFiles:q};t({x:r,y:s,width:d,height:l,payload:G})};return document.addEventListener("pointermove",e),{unsubscribe:()=>{document.removeEventListener("pointermove",e)},event:"contentlet-hovered"}}var R={changes:t=>x(t),"page-reload":t=>v(t),"request-bounds":t=>N(t),"iframe-scroll":t=>O(t),"contentlet-hovered":t=>L(t)},m=t=>({unsubscribe:()=>{},event:t});function E(){if(typeof window>"u"||window.parent===window||!window.location)return;let t=new URL(window.location.href),e=Object.values(c),n=t.searchParams.get("mode")??"EDIT_MODE",o=t.searchParams.get("language_id"),r=t.searchParams.get("personaId"),s=t.searchParams.get("variantName"),d=t.searchParams.get("experimentId"),l=t.searchParams.get("publishDate"),p=t.searchParams.get("dotCMSHost");return e.includes(n)||(n="EDIT_MODE"),{mode:n,languageId:o,persona:r,variantName:s,experimentId:d,publishDate:l,dotCMSHost:p}}function a(t,e){if(!E())return console.warn("UVE Subscription: Not running inside UVE"),m(t);let n=R[t];return n?n(e):(console.error(`UVE Subscription: Event ${t} not found`),m(t))}function w(t){i({action:"set-url",payload:{url:t||"/"}})}function i(t){window.parent.postMessage(t,"*")}function A(t){i({action:"edit-contentlet",payload:t})}function _(t){let{startLevel:e=1,depth:n=2}=t||{};i({action:"reorder-menu",payload:{startLevel:e,depth:n}})}function U(t,e){i({action:"init-inline-editing",payload:{type:t,data:e}})}function V(t){i({action:"set-bounds",payload:t})}function C(){let t=()=>{i({action:"scroll"})},e=()=>{i({action:"scroll-end"})};return window.addEventListener("scroll",t),window.addEventListener("scrollend",e),{destroyScrollHandler:()=>{window.removeEventListener("scroll",t),window.removeEventListener("scrollend",e)}}}function y(){document.querySelectorAll('[data-dot-object="contentlet"]').forEach(e=>{e.clientHeight||e.classList.add("empty-contentlet")})}function S(){let t=a("page-reload",()=>{window.location.reload()}),e=a("request-bounds",r=>{V(r)}),n=a("iframe-scroll",r=>{if(window.scrollY===0&&r==="up"||h()&&r==="down")return;let s=r==="up"?-120:120;window.scrollBy({left:0,top:s,behavior:"smooth"})}),o=a("contentlet-hovered",r=>{i({action:"set-contentlet",payload:r})});return{subscriptions:[t,e,n,o]}}function b(t){i({action:"client-ready",payload:t})}function D(){if(document.readyState==="complete")return k(),{destroyListenBlockEditorInlineEvent:()=>{}};let t=()=>{k()};return document.addEventListener("DOMContentLoaded",t),{destroyListenBlockEditorInlineEvent:()=>{document.removeEventListener("DOMContentLoaded",t)}}}var k=()=>{console.log("listenBlockEditorClick");let t=document.querySelectorAll("[data-block-editor-content]");t.length&&t.forEach(e=>{let{inode:n,language:o="1",contentType:r,fieldName:s,blockEditorContent:d}=e.dataset,l=JSON.parse(d||"");if(!n||!o||!r||!s){console.error("Missing data attributes for block editor inline editing."),console.warn("inode, language, contentType and fieldName are required.");return}e.classList.add("dotcms__inline-edit-field"),e.addEventListener("click",()=>{U("BLOCK_EDITOR",{inode:n,content:l,language:parseInt(o),fieldName:s,contentType:r})})})};var Y={createSubscription:a,editContentlet:A,reorderMenu:_,updateNavigation:w};window.dotUVE=Y;var j=E();j?.mode==="EDIT_MODE"&&(S(),C(),y(),b(),D());
+var c=(r=>(r.EDIT="EDIT_MODE",r.PREVIEW="PREVIEW_MODE",r.LIVE="LIVE",r.UNKNOWN="UNKNOWN",r))(c||{});function f(t){return t.map(e=>{let n=e.getBoundingClientRect(),o=Array.from(e.querySelectorAll('[data-dot-object="contentlet"]'));return{x:n.x,y:n.y,width:n.width,height:n.height,payload:JSON.stringify({container:T(e)}),contentlets:F(n,o)}})}function F(t,e){return e.map(n=>{let o=n.getBoundingClientRect();return{x:0,y:o.y-t.y,width:o.width,height:o.height,payload:JSON.stringify({container:n.dataset?.dotContainer?JSON.parse(n.dataset?.dotContainer):g(n),contentlet:{identifier:n.dataset?.dotIdentifier,title:n.dataset?.dotTitle,inode:n.dataset?.dotInode,contentType:n.dataset?.dotType}})}})}function T(t){return{acceptTypes:t.dataset?.dotAcceptTypes||"",identifier:t.dataset?.dotIdentifier||"",maxContentlets:t.dataset?.maxContentlets||"",uuid:t.dataset?.dotUuid||""}}function g(t){let e=t.closest('[data-dot-object="container"]');return e?T(e):(console.warn("No container found for the contentlet"),null)}function u(t){if(!t)return null;let e=t.querySelector('[data-dot-object="empty-content"]');return t?.dataset?.dotObject==="contentlet"||t?.dataset?.dotObject==="container"&&e||t?.dataset?.dotObject==="container"&&t.children.length===0?t:u(t?.parentElement)}function M(t){let e=t.querySelectorAll('[data-dot-object="vtl-file"]');return e.length?Array.from(e).map(n=>({inode:n.dataset?.dotInode,name:n.dataset?.dotUrl})):null}function h(){let t=document.documentElement.scrollHeight,e=window.innerHeight;return window.scrollY+e>=t}function x(t){let e=n=>{n.data.name==="uve-set-page-data"&&t(n.data.payload)};return window.addEventListener("message",e),{unsubscribe:()=>{window.removeEventListener("message",e)},event:"changes"}}function v(t){let e=n=>{n.data.name==="uve-reload-page"&&t()};return window.addEventListener("message",e),{unsubscribe:()=>{window.removeEventListener("message",e)},event:"page-reload"}}function N(t){let e=n=>{if(n.data.name==="uve-request-bounds"){let o=Array.from(document.querySelectorAll('[data-dot-object="container"]')),r=f(o);t(r)}};return window.addEventListener("message",e),{unsubscribe:()=>{window.removeEventListener("message",e)},event:"request-bounds"}}function O(t){let e=n=>{if(n.data.name==="uve-scroll-inside-iframe"){let o=n.data.direction;t(o)}};return window.addEventListener("message",e),{unsubscribe:()=>{window.removeEventListener("message",e)},event:"iframe-scroll"}}function L(t){let e=n=>{let o=u(n.target);if(!o)return;let{x:r,y:s,width:d,height:l}=o.getBoundingClientRect(),p=o.dataset?.dotObject==="container",H={identifier:"TEMP_EMPTY_CONTENTLET",title:"TEMP_EMPTY_CONTENTLET",contentType:"TEMP_EMPTY_CONTENTLET_TYPE",inode:"TEMPY_EMPTY_CONTENTLET_INODE",widgetTitle:"TEMP_EMPTY_CONTENTLET",baseType:"TEMP_EMPTY_CONTENTLET",onNumberOfPages:1},B={identifier:o.dataset?.dotIdentifier,title:o.dataset?.dotTitle,inode:o.dataset?.dotInode,contentType:o.dataset?.dotType,baseType:o.dataset?.dotBasetype,widgetTitle:o.dataset?.dotWidgetTitle,onNumberOfPages:o.dataset?.dotOnNumberOfPages},q=M(o),G={container:o.dataset?.dotContainer?JSON.parse(o.dataset?.dotContainer):g(o),contentlet:p?H:B,vtlFiles:q};t({x:r,y:s,width:d,height:l,payload:G})};return document.addEventListener("pointermove",e),{unsubscribe:()=>{document.removeEventListener("pointermove",e)},event:"contentlet-hovered"}}var R={changes:t=>x(t),"page-reload":t=>v(t),"request-bounds":t=>N(t),"iframe-scroll":t=>O(t),"contentlet-hovered":t=>L(t)},m=t=>({unsubscribe:()=>{},event:t});function E(){if(typeof window>"u"||window.parent===window||!window.location)return;let t=new URL(window.location.href),e=Object.values(c),n=t.searchParams.get("mode")??"EDIT_MODE",o=t.searchParams.get("language_id"),r=t.searchParams.get("personaId"),s=t.searchParams.get("variantName"),d=t.searchParams.get("experimentId"),l=t.searchParams.get("publishDate"),p=t.searchParams.get("dotCMSHost");return e.includes(n)||(n="EDIT_MODE"),{mode:n,languageId:o,persona:r,variantName:s,experimentId:d,publishDate:l,dotCMSHost:p}}function a(t,e){if(!E())return console.warn("UVE Subscription: Not running inside UVE"),m(t);let n=R[t];return n?n(e):(console.error(`UVE Subscription: Event ${t} not found`),m(t))}function w(t){i({action:"set-url",payload:{url:t||"/"}})}function i(t){window.parent.postMessage(t,"*")}function A(t){i({action:"edit-contentlet",payload:t})}function _(t){let{startLevel:e=1,depth:n=2}=t||{};i({action:"reorder-menu",payload:{startLevel:e,depth:n}})}function U(t,e){i({action:"init-inline-editing",payload:{type:t,data:e}})}function V(t){i({action:"set-bounds",payload:t})}function C(){let t=()=>{i({action:"scroll"})},e=()=>{i({action:"scroll-end"})};return window.addEventListener("scroll",t),window.addEventListener("scrollend",e),{destroyScrollHandler:()=>{window.removeEventListener("scroll",t),window.removeEventListener("scrollend",e)}}}function y(){document.querySelectorAll('[data-dot-object="contentlet"]').forEach(e=>{e.clientHeight||e.classList.add("empty-contentlet")})}function S(){let t=a("page-reload",()=>{window.location.reload()}),e=a("request-bounds",r=>{V(r)}),n=a("iframe-scroll",r=>{if(window.scrollY===0&&r==="up"||h()&&r==="down")return;let s=r==="up"?-120:120;window.scrollBy({left:0,top:s,behavior:"smooth"})}),o=a("contentlet-hovered",r=>{i({action:"set-contentlet",payload:r})});return{subscriptions:[t,e,n,o]}}function b(t){i({action:"client-ready",payload:t})}function D(){if(document.readyState==="complete")return k(),{destroyListenBlockEditorInlineEvent:()=>{}};let t=()=>{k()};return document.addEventListener("DOMContentLoaded",t),{destroyListenBlockEditorInlineEvent:()=>{document.removeEventListener("DOMContentLoaded",t)}}}var k=()=>{let t=document.querySelectorAll("[data-block-editor-content]");t.length&&t.forEach(e=>{let{inode:n,language:o="1",contentType:r,fieldName:s,blockEditorContent:d}=e.dataset,l=JSON.parse(d||"");if(!n||!o||!r||!s){console.error("Missing data attributes for block editor inline editing."),console.warn("inode, language, contentType and fieldName are required.");return}e.classList.add("dotcms__inline-edit-field"),e.addEventListener("click",()=>{U("BLOCK_EDITOR",{inode:n,content:l,language:parseInt(o),fieldName:s,contentType:r})})})};var Y={createSubscription:a,editContentlet:A,reorderMenu:_,updateNavigation:w};window.dotUVE=Y;var j=E();j?.mode==="EDIT_MODE"&&(S(),C(),y(),b(),D());
 })();
