@@ -1,6 +1,8 @@
 package com.dotcms;
 
 import com.dotcms.ai.api.OpenAIVisionAPIImplTest;
+import com.dotcms.contenttype.business.StoryBlockValidationTest;
+import com.dotcms.contenttype.test.StoryBlockUtilTest;
 import com.dotcms.jitsu.validators.AnalyticsValidatorUtilTest;
 import com.dotcms.junit.MainBaseSuite;
 import com.dotcms.rest.api.v1.drive.ContentDriveHelperContentletAPIComparisonTest;
@@ -51,7 +53,9 @@ import org.junit.runners.Suite;
         OpenAIVisionAPIImplTest.class,
         ContentDriveHelperContentletAPIComparisonTest.class,
         AppsAPIImplTest.class,
-        Task251103AddStylePropertiesColumnInMultiTreeTest.class
+        Task251103AddStylePropertiesColumnInMultiTreeTest.class,
+        StoryBlockValidationTest.class,
+        StoryBlockUtilTest.class
 })
 
 public class MainSuite3a {
