@@ -92,7 +92,7 @@ public class ESContentFactoryImplTest extends IntegrationTestBase {
 
         site = new SiteDataGen().nextPersisted();
     }
-    
+
     final ESContentFactoryImpl instance = new ESContentFactoryImpl();
 
     public static class TestCase {
@@ -139,11 +139,11 @@ public class ESContentFactoryImplTest extends IntegrationTestBase {
         for(Map<String,Object> r : dc.loadObjectResults()) {
             inodes.add((String)r.get("inode"));
         }
-        
+
         List<Contentlet> contentlets = instance.findContentlets(inodes);
-        
+
         Assert.assertEquals(inodes.size(), contentlets.size());
-        
+
         Set<String> inodesSet=new HashSet<>(inodes);
         for(Contentlet cc : contentlets) {
             Assert.assertTrue(inodesSet.remove(cc.getInode()));
@@ -152,10 +152,10 @@ public class ESContentFactoryImplTest extends IntegrationTestBase {
     }
 
     /**
-     * When calling the findContentlets(List<String> inodes), it should return the 
+     * When calling the findContentlets(List<String> inodes), it should return the
      * contentlets in the same order as they were asked for, meaning, if the list
      * of contentlets was sent in ordered by mod_date, then they should come out ordered
-     * by mod_date.  If you order them by random, then you get them out in that same 
+     * by mod_date.  If you order them by random, then you get them out in that same
      * random order
      * @throws Exception
      */
@@ -169,47 +169,47 @@ public class ESContentFactoryImplTest extends IntegrationTestBase {
         DotConnect dc=new DotConnect();
         dc.setSQL("select inode from contentlet order by mod_date desc");
         dc.setMaxRows(10);
-        
+
         List<String> inodesToOrderBy=new ArrayList<>();
         for(Map<String,Object> r : dc.loadObjectResults()) {
           inodesToOrderBy.add((String)r.get("inode"));
         }
 
-        // ten inodes 
+        // ten inodes
         assertTrue("We've added 10 contentlets", inodesToOrderBy.size()==10);
-        
+
         //load the cache with the 9th one and the 7th one
         instance.find(inodesToOrderBy.get(9));
         instance.find(inodesToOrderBy.get(7));
-        
-        
+
+
         List<Contentlet> contentlets = instance.findContentlets(inodesToOrderBy);
-        
+
         // this list should mirror the db query above
         Assert.assertEquals("our inodes and contentlet lists match",inodesToOrderBy.size(), contentlets.size());
-        
+
         // make sure the contentlets are in the same order
         for(int i=0;i<inodesToOrderBy.size();i++) {
           assertEquals(inodesToOrderBy.get(i),contentlets.get(i).getInode());
         }
-        
-        
+
+
         //randomize the order sent in
-        Collections.shuffle(inodesToOrderBy); 
+        Collections.shuffle(inodesToOrderBy);
         contentlets = instance.findContentlets(inodesToOrderBy);
-        
+
         // this list should mirror the shuffle above
         Assert.assertEquals("our inodes and contentlet lists match",inodesToOrderBy.size(), contentlets.size());
-        
+
         // make sure the contentlets are in the same order
         for(int i=0;i<inodesToOrderBy.size();i++) {
           assertEquals("content is in the right order",inodesToOrderBy.get(i),contentlets.get(i).getInode());
         }
-        
-        
-        
+
+
+
     }
-    
+
     @Test
     public void saveContentlets() throws Exception {
         try {
@@ -251,9 +251,9 @@ public class ESContentFactoryImplTest extends IntegrationTestBase {
                 .host(site)
                 .setProperty("title", "Bullish On America? Get On Board With Southwest Air")
                 .setProperty("urlTitle", "title")
-                .setProperty("body", "During the 1980s and 1990s Southwest Air (LUV) ")
+                .setProperty("body", TestDataUtils.BLOCK_EDITOR_DUMMY_CONTENT)
                 .setProperty("sysPublishDate", new Date()).nextPersisted();
-        
+
         new ContentletDataGen(blogContentType.id())
         .languageId(languageId)
         .host(site)
@@ -304,9 +304,9 @@ public class ESContentFactoryImplTest extends IntegrationTestBase {
         assertFalse(UtilMethods.isSet(searchHits.getHits()));
     }
 
-    
-    
-    
+
+
+
     /**
      * this method tests that if we are passing in asc or desc as a sort order (both invalid without a
      * specified field) then we do not blow up.
@@ -326,20 +326,20 @@ public class ESContentFactoryImplTest extends IntegrationTestBase {
         final SearchHits descendingHitsUpper = instance.indexSearch("*", 20, 0, "DESC");
         // we should have hits, as we are ignoring the invalid sort
         assert (descendingHitsUpper.getHits().length > 0);
-        
+
         final SearchHits ascHitsUpper = instance.indexSearch("*", 20, 0, "DESC");
         // we should have hits, as we are ignoring the invalid sort
-        assert (ascHitsUpper.getHits().length > 0);        
-        
-        
+        assert (ascHitsUpper.getHits().length > 0);
+
+
     }
 
-    
-    
-    
-    
-    
-    
+
+
+
+
+
+
     private float getMaxScore(SearchHit[] hits) {
         float maxScore = java.lang.Float.MIN_VALUE;
 
@@ -511,36 +511,36 @@ public class ESContentFactoryImplTest extends IntegrationTestBase {
 
     @Test
     public void test_findContentletByIdentifier() throws Exception {
-    
+
         final Language language1 = new UniqueLanguageDataGen().nextPersisted();
         final Language language2 = new UniqueLanguageDataGen().nextPersisted();
         final ContentType blogType = TestDataUtils.getBlogLikeContentType(site);
-        
+
 
         // create URL-Mapped content
         final Contentlet workingOneLanguage = new ContentletDataGen(blogType.id())
                 .languageId(language1.getId())
-                .setProperty("body", "myBody")
+                .setProperty("body", TestDataUtils.BLOCK_EDITOR_DUMMY_CONTENT)
                 .nextPersisted();
-        
+
         // create URL-Mapped content
         final Contentlet workingTwoLanguage = new ContentletDataGen(blogType.id())
                 .languageId(language1.getId())
-                .setProperty("body", "myBody")
+                .setProperty("body", TestDataUtils.BLOCK_EDITOR_DUMMY_CONTENT)
                 .nextPersisted();
-        
+
         // create URL-Mapped content
         final Contentlet publishedTwoLanguage2 = new ContentletDataGen(blogType.id())
                 .languageId(language2.getId())
-                .setProperty("body", "myBody")
+                .setProperty("body", TestDataUtils.BLOCK_EDITOR_DUMMY_CONTENT)
                 .setProperty("identifier", workingTwoLanguage.getIdentifier())
                 .nextPersisted();
-        
-        
+
+
         APILocator.getContentletAPI().publish(publishedTwoLanguage2, APILocator.systemUser(), false);
-        
-        
-        
+
+
+
         assertEquals("workingOneLanguage exists and is working", workingOneLanguage, instance.findContentletByIdentifier(workingOneLanguage.getIdentifier(), false, language1.getId()));
         assertNull("workingOneLanguage does not exist in 2nd language", instance.findContentletByIdentifier(workingOneLanguage.getIdentifier(), false, language2.getId()));
         assertNull("workingOneLanguage does not exist in live", instance.findContentletByIdentifier(workingOneLanguage.getIdentifier(), true, language1.getId()));
@@ -551,52 +551,52 @@ public class ESContentFactoryImplTest extends IntegrationTestBase {
         assertEquals("workingTwoLanguage exists in langauge2 and is live", publishedTwoLanguage2, instance.findContentletByIdentifier(workingTwoLanguage.getIdentifier(), true, language2.getId()));
 
     }
-    
+
     /**
      * This tests whether we are getting cached results from queries to elasticsearch and that these
      * results are invalidated when a new piece of content is checked in
-     * 
+     *
      * @throws Exception
      */
     @Test
     public void test_cached_es_query_response() throws Exception {
-        
+
         final Language language1 = new UniqueLanguageDataGen().nextPersisted();
 
         final ContentType blogType = TestDataUtils.getBlogLikeContentType(site);
-        
+
 
         assert(CacheLocator.getESQueryCache() !=null);
         final String liveQuery = "+baseType:1 +live:true" ;
         final String workingQuery = "+baseType:1 +live:false" ;
-        
+
 
         SearchHits hits = instance.indexSearch(liveQuery, 10, 0, null);
-        
+
         //assert we have results
         assertTrue(hits.getTotalHits().value > 0);
-        
+
         SearchHits hits2 = instance.indexSearch(liveQuery, 10, 0, null);
-        
+
         // hits and hits2 are the same object in memory (meaning, it came from cache)
         assertTrue(hits == hits2);
-        
+
 
         // checkin a new piece of content
         new ContentletDataGen(blogType.id())
                 .languageId(language1.getId())
-                .setProperty("body", "myBody")
+                .setProperty("body", TestDataUtils.BLOCK_EDITOR_DUMMY_CONTENT)
                 .nextPersisted();
-        
+
 
         SearchHits hits3 = instance.indexSearch(liveQuery, 10, 0, null);
-        
+
         // Checking in a new piece of content flushed the esQuerycache, we get new results
         assertTrue(hits != hits3);
         assertTrue(hits3.getTotalHits().value > 0);
     }
-    
-    
+
+
     /**
      * This test insures that we are taking all the parameters of the query into account when
      * building our cache key
@@ -607,12 +607,12 @@ public class ESContentFactoryImplTest extends IntegrationTestBase {
         final Language language1 = new UniqueLanguageDataGen().nextPersisted();
 
         final ContentType blogType = TestDataUtils.getBlogLikeContentType(site);
-        
+
         for(int i=0;i<10;i++) {
         // checkin a new piece of content
         Contentlet con = new ContentletDataGen(blogType.id())
                 .languageId(language1.getId())
-                .setProperty("body", "myBody")
+                .setProperty("body", TestDataUtils.BLOCK_EDITOR_DUMMY_CONTENT)
                 .nextPersisted();
             if(i % 2==0) {
                 APILocator.getContentletAPI().publish(con, APILocator.systemUser(), false);
@@ -620,28 +620,28 @@ public class ESContentFactoryImplTest extends IntegrationTestBase {
         }
         final String liveQuery = "+baseType:1 +live:true" ;
         final String workingQuery = "+baseType:1 +live:false" ;
-        
+
 
         //default
         SearchHits hits = instance.indexSearch(liveQuery, 5, 0, null);
-        
+
         // working index
         SearchHits hits1 = instance.indexSearch(workingQuery, 5, 0, null);
-        
+
         // different limit
         SearchHits hits2 = instance.indexSearch(liveQuery, 4, 0, null);
-        
+
         // different offset
         SearchHits hits3 = instance.indexSearch(liveQuery, 5, 1, null);
-        
+
         // different sort
         SearchHits hits4 = instance.indexSearch(liveQuery, 5, 0, "title desc");
-        
+
         // different sort direction
         SearchHits hits5 = instance.indexSearch(liveQuery, 5, 0, "title asc");
-        
-        
-        
+
+
+
         //assert we have results
         assertTrue(hits.getTotalHits().value > 0);
 
@@ -665,7 +665,7 @@ public class ESContentFactoryImplTest extends IntegrationTestBase {
         assertNotSame(hits, hits5);
         assertNotEquals(hits , hits5);
         assertTrue(hits5.getTotalHits().value > 0);
-        
+
 
     }
 
@@ -761,7 +761,7 @@ public class ESContentFactoryImplTest extends IntegrationTestBase {
     }
 
     /**
-     * Method to test: {@link ESContentFactoryImpl#findContentletByIdentifierAnyLanguage(String, boolean)} 
+     * Method to test: {@link ESContentFactoryImpl#findContentletByIdentifierAnyLanguage(String, boolean)}
      * Given Scenario: Get a contentlet given its identifier regardless of the language and its archived status
      * ExpectedResult: The method should return a contentlet
      *
@@ -896,7 +896,7 @@ public class ESContentFactoryImplTest extends IntegrationTestBase {
         assertNotNull(contentlet.getTitle());
 
         contentlet = instance.findInDb(contentlet.getInode()).get();
-        
+
         assertNull(contentlet.getMap().get(Contentlet.TITTLE_KEY));
     }
 
@@ -1131,7 +1131,7 @@ public class ESContentFactoryImplTest extends IntegrationTestBase {
         assertTrue(contentlets.stream().anyMatch(contentlet -> contentlet.getIdentifier()
                 .equals(contentletLanguage3.getIdentifier())));
     }
-    
+
     /**
      * Method to test: {@link ESContentFactoryImpl#findAllVersions(Identifier, boolean)}
      * When: The contentlet had several versions in different {@link Language} and some then are old versions
